Update crate references
