+++ bench-tps/Cargo.toml
@@ -12,6 +12,7 @@ clap = "2.32.0"
 rayon = "1.0.3"
 serde_json = "1.0.39"
 solana = { path = "../core", version = "0.13.0" }
+solana-client = { path = "../client", version = "0.13.0" }
 solana-drone = { path = "../drone", version = "0.13.0" }
 solana-logger = { path = "../logger", version = "0.13.0" }
 solana-metrics = { path = "../metrics", version = "0.13.0" }
