+++ core/Cargo.toml
@@ -39,13 +39,13 @@ nix = "0.13.0"
 rand = "0.6.5"
 rand_chacha = "0.1.1"
 rayon = "1.0.0"
-reqwest = "0.9.11"
 ring = "0.13.2"
 rocksdb = "0.11.0"
 serde = "1.0.89"
 serde_derive = "1.0.88"
 serde_json = "1.0.39"
 solana-budget-api = { path = "../programs/budget_api", version = "0.13.0" }
+solana-client = { path = "../client", version = "0.13.0" }
 solana-drone = { path = "../drone", version = "0.13.0" }
 solana-logger = { path = "../logger", version = "0.13.0" }
 solana-metrics = { path = "../metrics", version = "0.13.0" }
@@ -65,4 +65,3 @@ hex-literal = "0.1.3"
 matches = "0.1.6"
 solana-vote-program = { path = "../programs/vote", version = "0.13.0" }
 solana-budget-program = { path = "../programs/budget", version = "0.13.0" }
-
