+++ core/src/replicator.rs
@@ -3,19 +3,19 @@ use crate::blocktree::Blocktree;
 use crate::blocktree_processor;
 #[cfg(feature = "chacha")]
 use crate::chacha::{chacha_cbc_encrypt_ledger, CHACHA_BLOCK_SIZE};
-use crate::client::mk_client;
+use crate::cluster_client::mk_client;
 use crate::cluster_info::{ClusterInfo, Node};
 use crate::contact_info::ContactInfo;
 use crate::gossip_service::GossipService;
 use crate::result::Result;
-use crate::rpc_request::{RpcClient, RpcRequest, RpcRequestHandler};
 use crate::service::Service;
 use crate::storage_stage::{get_segment_from_entry, ENTRIES_PER_SEGMENT};
 use crate::streamer::BlobReceiver;
-use crate::thin_client::{retry_get_balance, ThinClient};
 use crate::window_service::WindowService;
 use rand::thread_rng;
 use rand::Rng;
+use solana_client::rpc_request::{RpcClient, RpcRequest, RpcRequestHandler};
+use solana_client::thin_client::{retry_get_balance, ThinClient};
 use solana_drone::drone::{request_airdrop_transaction, DRONE_PORT};
 use solana_sdk::genesis_block::GenesisBlock;
 use solana_sdk::hash::{Hash, Hasher};
