+++ wallet/Cargo.toml
@@ -18,6 +18,7 @@ log = "0.4.2"
 serde_json = "1.0.39"
 solana = { path = "../core", version = "0.13.0" }
 solana-budget-api = { path = "../programs/budget_api", version = "0.13.0" }
+solana-client = { path = "../client", version = "0.13.0" }
 solana-drone = { path = "../drone", version = "0.13.0" }
 solana-logger = { path = "../logger", version = "0.13.0" }
 solana-sdk = { path = "../sdk", version = "0.13.0" }
