+++ wallet/tests/request_airdrop.rs
@@ -1,5 +1,5 @@
-use solana::rpc_request::RpcClient;
-use solana::thin_client::new_fullnode;
+use solana::fullnode::new_fullnode_for_tests;
+use solana_client::rpc_request::RpcClient;
 use solana_drone::drone::run_local_drone;
 use solana_sdk::signature::KeypairUtil;
 use solana_wallet::wallet::{process_command, WalletCommand, WalletConfig};
@@ -8,7 +8,7 @@ use std::sync::mpsc::channel;
 
 #[test]
 fn test_wallet_request_airdrop() {
-    let (server, leader_data, alice, ledger_path) = new_fullnode();
+    let (server, leader_data, alice, ledger_path) = new_fullnode_for_tests();
     let (sender, receiver) = channel();
     run_local_drone(alice, sender);
     let drone_addr = receiver.recv().unwrap();
